Fix: Set Amount Limit & Use Double for amount value

+++ app/src/main/java/com/example/merabills/models/Payment.java
@@ -4,12 +4,12 @@ import com.example.merabills.enums.PaymentMode;
 
 public class Payment {
     public PaymentMode mode;
-    public float amount;
+    public double amount;
     public String provider;
     public String transactionReference;
 
     private Payment(){}
-    public Payment(float amount, PaymentMode mode, String provider, String transactionReference) {
+    public Payment(double amount, PaymentMode mode, String provider, String transactionReference) {
         this.mode = mode;
         this.amount = amount;
         this.provider = provider;
